+++ ClimateExplorer.Visualiser/Pages/RegionalAndGlobal.razor
@@ -4,6 +4,8 @@
 @using static ClimateExplorer.Visualiser.Shared.Collapsible;
 @using static ClimateExplorer.Visualiser.UiModel.SuggestedPresetLists;
 
+@inherits ChartablePage
+
 <PageTitle>@GetPageTitle()</PageTitle>
 
 <div style="flex-grow: 0">
