+++ ClimateExplorer.Visualiser/Pages/RegionalAndGlobal.razor.cs
@@ -1,42 +1,79 @@
-
+using Blazorise;
+using ClimateExplorer.Core.Infrastructure;
 using ClimateExplorer.Core.ViewModel;
+using ClimateExplorer.Visualiser.Services;
 using ClimateExplorer.Visualiser.Shared;
+using ClimateExplorer.Visualiser.UiLogic;
+using ClimateExplorer.Visualiser.UiModel;
+using DPBlazorMapLibrary;
 using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.Logging;
+using Microsoft.JSInterop;
 
 namespace ClimateExplorer.Visualiser.Pages;
-public partial class RegionalAndGlobal : IDisposable
+public partial class RegionalAndGlobal : ChartablePage
 {
-    [Inject] IDataService DataService { get; set; }
-    [Inject] ILogger<RegionalAndGlobal> Logger { get; set; }
-
-    IEnumerable<DataSetDefinitionViewModel> DataSetDefinitions { get; set; }
-
-    ChartView chartView;
+    public RegionalAndGlobal()
+    {
+        baseUrl = "/regionalandglobal";
+    }
+    
     Modal addDataSetModal { get; set; }
 
     protected override async Task OnInitializedAsync()
     {
-        if (DataService == null)
+        if (base.DataService == null)
         {
-            throw new NullReferenceException(nameof(DataService));
+            throw new NullReferenceException(nameof(base.DataService));
         }
-        DataSetDefinitions = (await DataService.GetDataSetDefinitions()).ToList();
+        DataSetDefinitions = (await base.DataService.GetDataSetDefinitions()).ToList();
     }
+
 
     public void Dispose()
     {
-        Logger.LogInformation("Instance " + _componentInstanceId + " disposing");
+        //Logger.LogInformation("Instance " + _componentInstanceId + " disposing");
         //   NavManager.LocationChanged -= HandleLocationChanged;
     }
+
+    async Task OnAddDataSet(DataSetLibraryEntry dle)
+    {
+        await base.chartView.OnAddDataSet(dle, DataSetDefinitions);
+    }
+
+    async Task OnChartPresetSelected(List<ChartSeriesDefinition> chartSeriesDefinitions)
+    {
+        await base.chartView.OnChartPresetSelected(chartSeriesDefinitions);
+    }
+
     string GetPageTitle()
     {
         //var locationText = SelectedLocation == null ? "" : " - " + SelectedLocation.Name;
 
         string title = $"ClimateExplorer";// {locationText}";
 
-        Logger.LogInformation("GetPageTitle() returning '" + title + "' NavigateTo");
+        base.Logger.LogInformation("GetPageTitle() returning '" + title + "' NavigateTo");
 
         return title;
+    }
+
+    private Task ShowAddDataSetModal()
+    {
+        return addDataSetModal.Show();
+    }
+
+    protected async Task OnDownloadDataClicked(DataDownloadPackage dataDownloadPackage)
+    {
+        var fileStream = Exporter.ExportChartData(Logger, new List<Location>(), dataDownloadPackage, NavManager.Uri.ToString());
+
+        var locationNames = dataDownloadPackage.ChartSeriesWithData.SelectMany(x => x.ChartSeries.SourceSeriesSpecifications).Select(x => x.LocationName).Where(x => x != null).Distinct().ToArray();
+
+        var fileName = locationNames.Any() ? string.Join("-", locationNames) + "-" : "";
+
+        fileName = $"Export-{fileName}-{dataDownloadPackage.BinGranularity}-{dataDownloadPackage.Bins.First().Label}-{dataDownloadPackage.Bins.Last().Label}.csv";
+
+        using var streamRef = new DotNetStreamReference(stream: fileStream);
+
+        await base.JsRuntime.InvokeVoidAsync("downloadFileFromStream", fileName, streamRef);
     }
 }
