+++ ClimateExplorer.Visualiser/Shared/SelectLocation.razor
@@ -36,7 +36,7 @@
                     TextField="@((item)=>item.Name)"
                     ValueField="@((item)=>item.Id)"
                     SelectedValueChanged="@OnLocationChange"
-                    DefaultItemText="Select Location"
+                    DefaultItemText="Select location"
                     Style="width: 100%"
                     />
            </Field>
