--- ClimateExplorer.Visualiser/UiModel/ChartViewModel.cs
@@ -1,5 +1,0 @@
-﻿namespace ClimateExplorer.Visualiser.UiModel;
-
-public class ChartViewModel
-{
-}
