+++ ClimateExplorer.Visualiser/UiModel/DataDownloadPackage.cs
@@ -6,4 +6,5 @@ public class DataDownloadPackage
 {
     public List<SeriesWithData> ChartSeriesWithData { get; set; }
     public BinIdentifier[] Bins { get; set; }
+    public BinGranularities BinGranularity { get; set; }
 }
